Update to compile on MacOS Tahoe 26.3 with latest packages on MacPorts. CMakeList.txt was updated to use the modern way to find Boost (see policy CMP0167). Argument to boost::posix_time::microseconds is now explicitly cast as a long to avoid a compilation error.

+++ src/main.cpp
@@ -64,7 +64,7 @@ int main(int argc, char *argv[])
 	latero::Tactograph dev;
 
 	std::cout << "Starting engines...\n";
-	latero::graphics::TactileEngine tEngine(&dev, boost::posix_time::microseconds(1E6/UpdateRateHz));
+	latero::graphics::TactileEngine tEngine(&dev, boost::posix_time::microseconds((long)(1E6/UpdateRateHz)));
 	latero::graphics::AudioEngine aEngine(&dev, boost::posix_time::milliseconds(30));
 	tEngine.Start();
 	// aEngine.Start(); // we don't need audio
